refac(ApplyTagsPopper): separate TagCheckbox The TagCheckbox component is now in it's own module.
